Make unconnected into a real module
